+++ main/src/ca/uwaterloo/flix/language/phase/MonomorphEnums.scala
@@ -21,7 +21,7 @@ import ca.uwaterloo.flix.language.ast.Ast.CaseSymUse
 import ca.uwaterloo.flix.language.ast.LoweredAst.{Expression, Pattern}
 import ca.uwaterloo.flix.language.ast.Type.eraseAliases
 import ca.uwaterloo.flix.language.ast.{Ast, LoweredAst, Name, Scheme, SourceLocation, Symbol, Type, TypeConstructor}
-import ca.uwaterloo.flix.language.phase.unification.Substitution
+import ca.uwaterloo.flix.language.phase.unification.{Substitution, TypeNormalization}
 import ca.uwaterloo.flix.util.InternalCompilerException
 
 import scala.collection.mutable
@@ -93,24 +93,41 @@ object MonomorphEnums {
 
     implicit val ctx: Context = new Context()
 
-    val defs = for ((sym, defn) <- root.defs) yield {
-      val spec0 = defn.spec
-      val spec = LoweredAst.Spec(
-        spec0.doc,
-        spec0.ann,
-        spec0.mod,
-        spec0.tparams,
-        spec0.fparams.map(visitFormalParam),
-        visitScheme(spec0.declaredScheme),
-        visitType(spec0.retTpe),
-        visitType(spec0.pur),
-        spec0.tconstrs, spec0.loc
-      )
-      val impl = LoweredAst.Impl(visitExp(defn.impl.exp), visitScheme(defn.impl.inferredScheme))
-      (sym, LoweredAst.Def(sym, spec, impl))
-    }
+    val defs = root.defs.view.mapValues(visitDef).toMap
+    val enums = ctx.specializedEnums.toMap
+    root.copy(defs = defs, enums = enums)
+  }
 
-    root.copy(defs = defs, enums = ctx.specializedEnums.toMap)
+  /**
+    * Returns a [[LoweredAst.Def]] with specialized enums and without aliases in its types.
+    */
+  private def visitDef(defn: LoweredAst.Def)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): LoweredAst.Def = defn match {
+    case LoweredAst.Def(sym, spec, impl) =>
+      val s = visitSpec(spec)
+      val i = visitImpl(impl)
+      LoweredAst.Def(sym, s, i)
+  }
+
+  /**
+    * Returns a [[LoweredAst.Spec]] with specialized enums and without aliases in its types.
+    */
+  private def visitSpec(spec: LoweredAst.Spec)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): LoweredAst.Spec = spec match {
+    case LoweredAst.Spec(doc, ann, mod, tparams, fparams, declaredScheme, retTpe, pur, tconstrs, loc) =>
+      val fs = fparams.map(visitFormalParam)
+      val ds = visitScheme(declaredScheme)
+      val rt = visitType(retTpe)
+      val p = visitType(pur)
+      LoweredAst.Spec(doc, ann, mod, tparams, fs, ds, rt, p, tconstrs, loc)
+  }
+
+  /**
+    * Returns a [[LoweredAst.Impl]] with specialized enums and without aliases in its types.
+    */
+  private def visitImpl(impl: LoweredAst.Impl)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): LoweredAst.Impl = impl match {
+    case LoweredAst.Impl(exp, inferredScheme) =>
+      val e = visitExp(exp)
+      val is = visitScheme(inferredScheme)
+      LoweredAst.Impl(e, is)
   }
 
   /**
@@ -484,19 +501,20 @@ object MonomorphEnums {
   /**
     * Returns a formal param with specialized enums in its type and no aliases.
     */
-  private def visitFormalParam(p: LoweredAst.FormalParam)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): LoweredAst.FormalParam = {
-    val LoweredAst.FormalParam(sym, mod, tpe, src, loc) = p
-    val t = visitType(tpe)
-    LoweredAst.FormalParam(sym, mod, t, src, loc)
+  private def visitFormalParam(p: LoweredAst.FormalParam)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): LoweredAst.FormalParam = p match {
+    case LoweredAst.FormalParam(sym, mod, tpe, src, loc) =>
+      val t = visitType(tpe)
+      LoweredAst.FormalParam(sym, mod, t, src, loc)
   }
 
   /**
     * Returns a scheme with specialized enums in its base and no aliases.
     */
-  private def visitScheme(sc: Scheme)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): Scheme = {
-    val Scheme(quantifiers, tconstrs, econstrs, base) = sc
-    val b = visitType(base)
-    Scheme(quantifiers, tconstrs, econstrs, b)
+  private def visitScheme(sc: Scheme)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): Scheme = sc match {
+    case Scheme(quantifiers, tconstrs, econstrs, base) =>
+      // Since the types are expected to be specialized, all except base should be "unused"/empty
+      val b = visitType(base)
+      Scheme(quantifiers, tconstrs, econstrs, b)
   }
 
   /**
@@ -510,7 +528,7 @@ object MonomorphEnums {
     * - may be un-normalized
     */
   private def specializeEnum(sym: Symbol.EnumSym, args0: List[Type], loc: SourceLocation)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): Symbol.EnumSym = {
-    val args = args0.map(eraseAliases).map(normalizeType)
+    val args = args0.map(eraseAliases).map(TypeNormalization.normalizeType)
     // assemble enum type (e.g. `List[Int32]`)
     val tpe = Type.mkEnum(sym, args, loc)
     // reuse specialization if possible
@@ -559,140 +577,6 @@ object MonomorphEnums {
   private def specializeCaseSymUse(sym: CaseSymUse, args: List[Type], loc: SourceLocation)(implicit ctx: Context, root: LoweredAst.Root, flix: Flix): CaseSymUse = {
     val freshEnumSym = specializeEnum(sym.sym.enumSym, args, loc)
     Ast.CaseSymUse(new Symbol.CaseSym(freshEnumSym, sym.sym.name, sym.sym.loc), sym.loc)
-  }
-
-  /**
-    * Converts a type into an equivalent type in normalized form, which will be
-    * the same for all other equivalent types.
-    *
-    * Returns a type where
-    * - Formulas in types have been fully evaluated (and ordered in the case of sets)
-    * - Types involving rows have been sorted alphabetically (respecting duplicate label ordering)
-    * - The assumptions still hold
-    *
-    * Assumes that
-    * - `tpe` is ground (no type variables)
-    * - `tpe` has no aliases
-    * - `tpe` has no associated types
-    */
-  private def normalizeType(tpe: Type): Type = tpe match {
-    case Type.Var(sym, loc) =>
-      throw InternalCompilerException(s"Unexpected type var '$sym'", loc)
-    case Type.Cst(_, _) =>
-      tpe
-    case Type.Apply(tpe1, tpe2, applyLoc) =>
-      val t1 = normalizeType(tpe1)
-      val t2 = normalizeType(tpe2)
-      t1 match {
-        // Simplify effect set equations.
-        case Type.Cst(TypeConstructor.Complement, _) => t2 match {
-          case Type.Pure => Type.EffUniv
-          case Type.EffUniv => Type.Pure
-          case _ => throw InternalCompilerException(s"Unexpected non-simple effect $tpe", applyLoc)
-        }
-        case Type.Apply(Type.Cst(TypeConstructor.Union, _), x, _) =>
-          (x, t2) match {
-            case (Type.Pure, Type.Pure) => Type.Pure
-            case (Type.Pure, Type.EffUniv) => Type.EffUniv
-            case (Type.EffUniv, Type.Pure) => Type.EffUniv
-            case (Type.EffUniv, Type.EffUniv) => Type.EffUniv
-            case _ => throw InternalCompilerException(s"Unexpected non-simple effect $tpe", applyLoc)
-          }
-        case Type.Apply(Type.Cst(TypeConstructor.Intersection, _), x, _) =>
-          (x, t2) match {
-            case (Type.Pure, Type.Pure) => Type.Pure
-            case (Type.Pure, Type.EffUniv) => Type.Pure
-            case (Type.EffUniv, Type.Pure) => Type.Pure
-            case (Type.EffUniv, Type.EffUniv) => Type.EffUniv
-            case _ => throw InternalCompilerException(s"Unexpected non-simple effect $tpe", applyLoc)
-          }
-
-        // Simplify boolean equations.
-        // TODO EFF-MIGRATION
-        case Type.Cst(TypeConstructor.Not, _) |
-             Type.Apply(Type.Cst(TypeConstructor.And, _), _, _) |
-             Type.Apply(Type.Cst(TypeConstructor.Or, _), _, _) =>
-          throw InternalCompilerException(s"Unexpected Not/And/Or in formula $tpe", applyLoc)
-
-        // Simplify set expressions
-        case Type.Cst(TypeConstructor.CaseComplement(enumSym), _) => t2 match {
-          case Type.Cst(TypeConstructor.CaseSet(syms, _), loc) =>
-            Type.Cst(TypeConstructor.CaseSet(enumSym.universe.diff(syms), enumSym), loc)
-          case _ => throw InternalCompilerException(s"Unexpected non-simple case set formula $tpe", applyLoc)
-        }
-        case Type.Apply(Type.Cst(TypeConstructor.CaseIntersection(enumSym), _), x, loc) =>
-          (x, t2) match {
-            case (Type.Cst(TypeConstructor.CaseSet(syms1, _), _), Type.Cst(TypeConstructor.CaseSet(syms2, _), _)) =>
-              Type.Cst(TypeConstructor.CaseSet(syms1.intersect(syms2), enumSym), loc)
-            case _ => throw InternalCompilerException(s"Unexpected non-simple case set formula $tpe", applyLoc)
-          }
-        case Type.Apply(Type.Cst(TypeConstructor.CaseUnion(enumSym), _), x, loc) =>
-          (x, t2) match {
-            case (Type.Cst(TypeConstructor.CaseSet(syms1, _), _), Type.Cst(TypeConstructor.CaseSet(syms2, _), _)) =>
-              Type.Cst(TypeConstructor.CaseSet(syms1.union(syms2), enumSym), loc)
-            case _ => throw InternalCompilerException(s"Unexpected non-simple case set formula $tpe", applyLoc)
-          }
-
-        // Sort record row fields
-        case Type.Apply(Type.Cst(TypeConstructor.RecordRowExtend(field), _), fieldType, _) =>
-          insertRecordField(field, fieldType, t2, applyLoc)
-
-        // Sort schema row fields
-        case Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(pred), _), predType, _) =>
-          insertSchemaPred(pred, predType, t2, applyLoc)
-
-        // Else just apply
-        case x => Type.Apply(x, t2, applyLoc)
-      }
-    case Type.Alias(cst, _, _, loc) =>
-      throw InternalCompilerException(s"Unexpected type alias: '${cst.sym}'", loc)
-    case Type.AssocType(cst, _, _, loc) =>
-      throw InternalCompilerException(s"Unexpected associated type: '${cst.sym}'", loc)
-  }
-
-  /**
-    * Inserts the given field into `rest` in its ordered position, assuming that
-    * `rest` is already ordered. This, together with [[normalizeType]]
-    * effectively implements insertion sort.
-    */
-  private def insertRecordField(field: Name.Field, fieldType: Type, rest: Type, loc: SourceLocation): Type = rest match {
-    // empty rest, create the singleton record row
-    case Type.Cst(TypeConstructor.RecordRowEmpty, emptyLoc) =>
-      Type.mkRecordRowExtend(field, fieldType, Type.mkRecordRowEmpty(emptyLoc), loc)
-    // the current field should be before the next field and since
-    // - we insert from the left, one by one
-    // - rest is ordered
-    // we can return the current field with the rest
-    case Type.Apply(Type.Apply(Type.Cst(TypeConstructor.RecordRowExtend(field1), _), _, _), _, _) if field.name <= field1.name =>
-      Type.mkRecordRowExtend(field, fieldType, rest, loc)
-    // The current field should be after the next field, so we swap and continue recursively
-    case Type.Apply(Type.Apply(Type.Cst(TypeConstructor.RecordRowExtend(field1), field1Loc), field1Type, field1TypeLoc), rest1, rest1Loc) =>
-      val tail = insertRecordField(field, fieldType, rest1, loc)
-      Type.Apply(Type.Apply(Type.Cst(TypeConstructor.RecordRowExtend(field1), field1Loc), field1Type, field1TypeLoc), tail, rest1Loc)
-    case other => throw InternalCompilerException(s"Unexpected record rest: '$other'", rest.loc)
-  }
-
-  /**
-    * Inserts the given predicate into `rest` in its ordered position, assuming that
-    * `rest` is already ordered. This, together with [[normalizeType]]
-    * effectively implements insertion sort.
-    */
-  private def insertSchemaPred(pred: Name.Pred, predType: Type, rest: Type, loc: SourceLocation): Type = rest match {
-    // empty rest, create the singleton schema row
-    case Type.Cst(TypeConstructor.SchemaRowEmpty, _) =>
-      Type.mkSchemaRowExtend(pred, predType, rest, loc)
-    // the current pred should be before the next pred and since
-    // - we insert from the left, one by one
-    // - rest is ordered
-    // we can return the current pred with the rest
-    case Type.Apply(Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(pred1), _), _, _), _, _) if pred.name <= pred1.name =>
-      Type.mkSchemaRowExtend(pred, predType, rest, loc)
-    // The current pred should be after the next pred, so we swap and continue recursively
-    case Type.Apply(Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(pred1), pred1Loc), pred1Type, pred1TypeLoc), rest1, rest1Loc) =>
-      val rest2 = insertSchemaPred(pred, predType, rest1, loc)
-      Type.Apply(Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(pred1), pred1Loc), pred1Type, pred1TypeLoc), rest2, rest1Loc)
-    case other =>
-      throw InternalCompilerException(s"Unexpected schema rest: '$other'", rest.loc)
   }
 
 }
